# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LexiVerse Scholarly Research API",
    "description": "Integration portal for the LexiVerse research engine. Provides access to AI-powered scripture analysis and linguistic synthesis.",
    "version": "1.0.0",
    "contact": {
      "name": "LexiVerse Support",
      "url": "https://lexiverse.app/help"
    }
  },
  "servers": [
    {
      "url": "https://lexiverse.app/api/v1",
      "description": "Production Scholarly Hub"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "LexiVerse Key (lv_...)"
      }
    },
    "schemas": {
      "ResearchRequest": {
        "type": "object",
        "required": ["term"],
        "properties": {
          "term": {
            "type": "string",
            "description": "The scripture term or reference to research (e.g., 'G3056' or 'John 1:1')."
          },
          "researchContext": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Optional excerpts from research papers to ground the AI in specific context (RAG)."
          }
        }
      },
      "ResearchResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "originalWord": { "type": "string" },
              "transliteration": { "type": "string" },
              "aiInsights": { "type": "string" },
              "verseUsages": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "text": { "type": "string" },
                    "url": { "type": "string" }
                  }
                }
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "tier": { "type": "string" },
              "remaining": { "type": "number" }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/research": {
      "post": {
        "summary": "Execute Scholarly Research",
        "description": "Submits a term for analysis. Returns linguistic data, commentary synthesis, and bibliographic references.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResearchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Analysis",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResearchResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - Missing Bearer Token" },
          "403": { "description": "Forbidden - Invalid or Revoked Key" },
          "429": { "description": "Tier Limit Exceeded" }
        }
      }
    }
  }
}